+++ src/assets/lang/en.json
@@ -568,8 +568,7 @@
     "disclaimer": "**NOTE:** This is for technical feedback only and will be shared with our app developers.\n\nIf you are feeling sick or have general questions about COVID-19, please do not use this feedback option and call the COVID Alert NY Hotline instead at [1-833-227-5045](tel:18332275045)."
   },
   "ageCheck": {
-    "intro": "Are you over 18?",
-    "parentConsent": "Did your parent or guardian confirm that you can use this app?",
-    "underAge": "Sorry... You cannot use this app without parent or guardian confirmation."
+    "intro": "I am over 18, or my parent or guardian has confirmed that I can use this app.",
+    "confirm": "I Confirm"
   }
 }
