+++ src/components/views/onboarding/age-check/index.tsx
@@ -1,5 +1,3 @@
 import {AgeCheck} from './age-check';
-import {ParentConsent} from './parent-consent';
-import {UnderAge} from './under-age';
 
-export {AgeCheck, ParentConsent, UnderAge};
+export {AgeCheck};
